Extract packet handling to new module
